MO-9233: QLibraryInfo is able to use environment variable for base path
The variable name is MOCHA_PLUGIN_MODULE_PATH and it should contain path
to the plugin directory. For example:
/usr/OFX/Plugins/BorisFX/MochaV6/SharedResources/lib

This path helps Qt to locate all required resources listed in the
qt.conf file. If this variable is not set,
QCoreApplication::applicationDirPath() or QDir::currentPath() will be
used. They might contain wrong path for plugin i.e. host app path.